requirements.txt
alembic==1.10.2
apispec==4.7.1
click==7.1.2
ComboJSONAPI==1.1.1
dnspython==2.3.0
email-validator==1.1.2
Flask==1.1.4
Flask-Admin==1.5.8
Flask-COMBO-JSONAPI==1.0.7
Flask-Login==0.5.0
Flask-Migrate==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
idna==3.4
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.2.4
MarkupSafe==2.0.1
marshmallow==3.2.1
marshmallow-jsonapi==0.24.0
psycopg2-binary==2.9.5
python-dotenv==0.17.1
PyYAML==5.4.1
simplejson==3.18.4
SQLAlchemy==1.3.24
typing_extensions==4.5.0
Werkzeug==0.16.1
WTForms==2.3.3
